+++ CHANGES.md
@@ -85,6 +85,7 @@ Changes in 13.0
 - Use DBInsert() & DBUpdate() functions (Moodle plugin) in ImportUsers.fnc.php, Student.php, User.php, CreateParents.php, Calendar.php, Courses.php
 - Update Russian and Ukrainian translations in rosariosis.po & help.po, thanks to Aleksey
 - Remove Portal Notes integration code (Moodle plugin) in PortalNotes.php & functions.php
+- Fix Secondary Teacher dropdown options when Moodle course in Courses.php
 
 Changes in 12.9.4
 -----------------